+++ tests/Unit/Schema/Source/SchemaStitcherTest.php
@@ -16,8 +16,8 @@ class SchemaStitcherTest extends TestCase
     {
         parent::setUp();
 
-        // @phpstan-ignore-next-line using the Safe variant crashes PHPStan
-        exec('mkdir --parents ' . self::SCHEMA_PATH);
+        // @phpstan-ignore-next-line using the Safe variant crashes PHPStan (uses the short `-p` because `--parent` is not available on macOS)
+        exec('mkdir -p ' . self::SCHEMA_PATH);
     }
 
     protected function tearDown(): void