+++ tests/Utils/Mutations/Upload.php
@@ -6,6 +6,9 @@ use Illuminate\Http\UploadedFile;
 
 final class Upload
 {
+    /**
+     * @param  array<string, mixed> $args
+     */
     public function __invoke($root, array $args): bool
     {
         return isset($args['file'])